requirements.txt
absl-py==2.1.0
accelerate==0.31.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
albumentations==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
basicsr==1.4.2
beautifulsoup4==4.12.3
blessed==1.20.0
blinker==1.8.2
boto3==1.34.138
botocore==1.34.138
braceexpand==0.1.7
cachetools==5.3.3
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cmake==3.29.6
colorama==0.4.6
colorlog==6.8.2
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.2.1
croniter==1.3.15
cupy-cuda11x==13.3.0
cycler==0.12.1
Cython==3.0.11
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dateutils==0.6.12
decorator==5.1.1
deepdiff==7.0.1
diff-gaussian-rasterization @ git+https://github.com/graphdeco-inria/diff-gaussian-rasterization.git

dill==0.3.8
dm-tree==0.1.8
docker-pycreds==0.4.0
editor==1.6.6
einops==0.8.0
embreex==2.17.7.post4
envlight @ git+https://github.com/ashawkey/envlight.git@05b5851e854429d72ecaf5b206ed64ce55fae677
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.88.0
fastjsonschema==2.20.0
fastrlock==0.8.2
ffmpy==0.3.2
filelock==3.14.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2023.12.2
ftfy==6.2.0
future==1.0.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2==2.1.5
google-pasta==0.2.0
GPUtil==1.4.0
GPy==1.13.2
gpytorch==1.12
gradio==3.50.2
gradio_client==0.6.1
grpcio==1.64.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
importlib_metadata==8.0.0
importlib_resources==6.4.0
inquirer==3.3.0
invisible-watermark==0.2.0
iopath==0.1.10
ipython==8.18.1
ipywidgets==8.1.5
itsdangerous==2.2.0
jaxtyping==0.2.19
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
keras==3.4.1
kiui @ git+https://github.com/ashawkey/kiuikit@ad4cbffb304a0053172c833cbbd30558a4087b6e
kiwisolver==1.4.5
kornia==0.7.3
kornia_rs==0.1.4
lazy_loader==0.4
libclang==18.1.1
libigl==2.5.1
lightning==2.0.0
lightning-cloud==0.5.70
lightning-utilities==0.11.3.post0
linear-operator==0.5.3
lit==18.1.8
lmdb==1.5.1
loguru==0.7.2
lpips==0.1.4
lxml==5.2.2
mapbox-earcut==1.0.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.8
nbformat==5.10.4
nerfacc @ git+https://github.com/KAIR-BAIR/nerfacc.git@d84cdf3afd7dcfc42150e0f0506db58a5ce62812
nest-asyncio==1.6.0
networkx==3.3
nflows==0.14
ninja==1.11.1.1
normflows==1.7.3
numpy==1.24.1
nvdiffrast @ git+https://github.com/NVlabs/nvdiffrast.git@c5caf7bdb8a2448acc491a9faa47753972edd380
objaverse==0.1.7
objprint==0.2.3
omegaconf==2.3.0
open3d==0.18.0
open_clip_torch==2.26.1
opencv-contrib-python==4.10.0.84
opencv-python==4.8.1.78
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
optree==0.12.1
ordered-set==4.1.0
orjson==3.10.6
packaging==24.1
pandas==2.2.2
paramz==0.9.6
parso==0.8.4
peft==0.7.0
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.2
plotly==5.24.1
plyfile==0.9.3
portalocker==2.10.0
prettytable==3.10.0
prompt_toolkit==3.0.48
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==16.1.0
pycollada==0.8
pydantic==1.7.4
pydantic_core==2.20.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.8.0
PyMCubes==0.1.4
pymeshlab==2023.12.post1
pyparsing==3.1.2
pyquaternion==0.9.9
pyre-extensions==0.0.29
pysdf==0.1.9
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytorch-lightning==1.9.1
pytorch-msssim==1.0.0
pytorch-ssim==0.1
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
qudida==0.0.4
readchar==4.1.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
retrying==1.3.4
rich==13.7.1
rpds-py==0.18.1
Rtree==1.2.0
runs==1.2.2
s3transfer==0.10.2
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.12.0
semantic-version==2.10.0
sentry-sdk==2.7.1
setproctitle==1.3.3
shapely==2.0.4
shellingham==1.5.4
simple-knn @ git+https://github.com/ashawkey/simple-knn.git
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.22.0
starsessions==1.3.0
streamlit==1.36.0
streamlit-drawable-canvas==0.9.3
svg.path==6.3
sympy==1.12
taichi==1.5.0
tb-nightly==2.18.0a20240707
tenacity==8.5.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-probability==0.24.0
termcolor==2.4.0
test-tube==0.7.5
tf_keras==2.17.0
threadpoolctl==3.5.0
tifffile==2024.7.2
timm==1.0.7
tinycudann @ git+https://github.com/NVlabs/tiny-cuda-nn.git@4892659e015b12bd2fce446b8a8e33e646b21746#subdirectory=bindings/torch
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.0.1
torch-cluster==1.6.3+pt20cu117
torch-fidelity==0.3.0
torch-geometric==2.6.1
torch-scatter==2.1.2+pt20cu117
torch-sparse==0.6.18+pt20cu117
torch-spline-conv==1.2.2+pt20cu117
torchmetrics==1.4.0.post0
torchvision==0.15.2
tornado==6.4.1
tqdm==4.66.1
traitlets==5.14.3
transformers==4.30.2
trimesh==4.4.1
triton==2.0.0
typeguard==2.13.3
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing-inspect==0.8.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
uvicorn==0.30.1
varname==0.13.1
vhacdx==0.0.8.post1
wandb==0.17.3
warp-lang==0.10.1
watchdog==4.0.1
wcwidth==0.2.13
webdataset==0.2.86
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.0.3
widgetsnbextension==4.0.13
wrapt==1.16.0
xatlas==0.0.9
xformers==0.0.20
xmod==1.8.1
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zipp==3.19.2
